channel_service/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.12.0
psycopg2-binary==2.9.3
gspread==5.3.2
xmltodict==0.13.0
celery==5.2.6
django-celery-results==2.3.1
redis==4.3.1
